trying to refactor dataframe

+++ dreem_tools/run.py
@@ -7,24 +7,59 @@ import os
 import re
 import pickle
 
-from dreem import bit_vector
 from dreem_tools import logger, util
+from dreem_tools.util import load
 
 
-def process_seq_csv(path):
-    df = pd.read_csv(path)
-    seq = "".join(list(df["nuc"]))
-    ss = "".join(list(df["struc"]))
-    data = ";".join([str(x) for x in list(df["mismatches"])])
-    return (seq, ss, data)
+def parse_demultiplex_log(lines):
+    data = []
+    for l in lines:
+        if l.startswith("#"):
+            continue
+        spl = l.split()
+        # check if last row has number of reads
+        try:
+            num = float(spl[-1])
+        except:
+            continue
+        data.append(spl)
+    df = pd.DataFrame(data, columns="id,tag,count".split(","))
+    df["count"] = pd.to_numeric(df["count"])
+    df.to_csv("demultiplex.csv", index=False)
+    return df
 
 
-def load(pathname):
-    """Safe loader for binary pickle files"""
-    fh = open(pathname, "rb")
-    data = pickle.load(fh)
-    fh.close()
-    return data
+def generate_barcode_file(df):
+    """
+    generates a .fa file for novobarcode
+    :param df: a dataframe with that contains informmation of barcode sequences
+    :return: None
+    """
+    log = logger.get_logger("RUN")
+    expects = ["barcode", "barcode_seq", "construct"]
+    for e in expects:
+        if e not in df:
+            raise ValueError(f"{e} column is required in the csv")
+    s = "Distance\t4\nFormat\t5\n"
+    seen = []
+    warning = False
+    for i, row in df.iterrows():
+        if row["barcode"] in seen:
+            log.warning(
+                f"{row['barcode']} has been used more than once this may be an issue"
+            )
+            warning = True
+            continue
+        s += f"{row['barcode']}\t{row['barcode_seq']}\n"
+        seen.append(row["barcode"])
+    log.info(f"{len(seen)} unique barcodes found from csv file")
+    if not warning:
+        log.info("no barcode conflicts detected")
+    f = open("rtb_barcodes.fa", "w")
+    f.write(s)
+    f.close()
+
+
 
 
 @click.group()
@@ -33,36 +68,52 @@ def cli():
 
 
 @cli.command()
+@click.argument("run_name")
+@click.option("-d", "--dir", default="/Users/jyesselm/BaseSpace")
+def download(run_name, dir):
+    os.chdir(dir)
+    os.makedirs(run_name, exist_ok=True)
+    os.chdir(run_name)
+    os.system(f"bs download project --name {run_name}")
+
+
+@cli.command()
 @click.argument("csv")
-def demultiplex(csv):
+@click.option("--debug", is_flag=True)
+def demultiplex(csv, debug):
     """
     demultiplexes paired fastq files given 3' end barcodes
     """
-    s = "Distance\t4\nFormat\t5\n"
+    log = logger.setup_applevel_logger(file_name='demultplex.log')
+    log.info("preparing rtb_barcodes.fa file for demultiplexing")
     df = pd.read_csv(csv)
-    seen = []
-    for i, row in df.iterrows():
-        if row["barcode"] in seen:
-            continue
-        s += f"{row['barcode']}\t{row['barcode_seq']}\n"
-        seen.append(row["barcode"])
-    f = open("rtb_barcodes.fa", "w")
-    f.write(s)
-    f.close()
+    generate_barcode_file(df)
     r1_fastq = glob.glob("*R1*")[0]
     r2_fastq = glob.glob("*R2*")[0]
+    log.info(f"renaming {r1_fastq} -> test_S1_L001_R1_001.fastq")
+    log.info(f"renaming {r2_fastq} -> test_S1_L001_R2_001.fastq")
     shutil.move(r1_fastq, "test_S1_L001_R1_001.fastq")
     shutil.move(r2_fastq, "test_S1_L001_R2_001.fastq")
-    subprocess.call(
-            "novobarcode -b rtb_barcodes.fa -f test_S1_L001_R1_001.fastq test_S1_L001_R2_001.fastq",
-            shell=True,
+    output = subprocess.check_output(
+        "novobarcode -b rtb_barcodes.fa -f test_S1_L001_R1_001.fastq test_S1_L001_R2_001.fastq",
+        shell=True,
+    )
+    output = output.decode("UTF-8")
+    log.info(f"output from novobarcode:\n{output}")
+    lines = output.split("\n")
+    df_demult = parse_demultiplex_log(lines)
+    log.info(f"total number of reads: {df_demult['count'].sum()}")
+    log.info(
+        f"total number of data reads: "
+        f"{df_demult['count'].sum() - df_demult.iloc[-1]['count']}"
     )
     if not os.path.isdir("NC"):
-        print("nothing was generated STOPING NOW!")
+        log.error("nothing was generated STOPING NOW!")
         exit()
-    os.remove("test_S1_L001_R1_001.fastq")
-    os.remove("test_S1_L001_R2_001.fastq")
-    shutil.rmtree("NC")
+    if not debug:
+        os.remove("test_S1_L001_R1_001.fastq")
+        os.remove("test_S1_L001_R2_001.fastq")
+        shutil.rmtree("NC")
 
 
 @cli.command()
@@ -72,21 +123,22 @@ def demultiplex(csv):
 @click.option("--move-plots", is_flag=True)
 def runmulti(csv, data_dir, seq_path, move_plots):
     # TODO write infomation about the run here
-    log = logger.setup_applevel_logger()
+    log = logger.setup_applevel_logger(file_name='run_multi.log')
+    log.info("creating processed/ all dreem runs will go here")
+    log.info("creating analysis/ all finalized analysis will go here")
     os.makedirs("processed", exist_ok=True)
     os.makedirs("analysis", exist_ok=True)
     df = pd.read_csv(csv)
     # catch old column naming scheme
     if "name" in df:
-        df = df.rename({"name" : "construct"},  axis='columns')
+        df = df.rename({"name": "construct"}, axis="columns")
     if "type" in df:
-        df = df.rename({"type" : "data_type"},  axis='columns')
+        df = df.rename({"type": "data_type"}, axis="columns")
     os.chdir("processed")
     if move_plots:
         os.makedirs("plots", exist_ok=True)
     dfs = []
     for i, row in df.iterrows():
-        print(row['construct'], row['code'], row['data_type'])
         dir_name = row["construct"] + "_" + row["code"] + "_" + row["data_type"]
         os.makedirs(dir_name, exist_ok=True)
         os.chdir(dir_name)
@@ -108,25 +160,6 @@ def runmulti(csv, data_dir, seq_path, move_plots):
         os.chdir("..")
     df_sum_final = pd.concat(dfs)
     df_sum_final.to_json(f"../analysis/summary.json", orient="records")
-
-
-
-@cli.command()
-@click.argument("pickle_files", nargs=-1)
-@click.option("-o", "--output", default="output.p")
-def merge(pickle_files, output):
-    log = logger.setup_applevel_logger()
-    log.info(f"{len(pickle_files)} pickle files supplied")
-    merged_mh = None
-    for pf in pickle_files:
-        if merged_mh is None:
-            merged_mh = load(pf)
-            continue
-        mh = load(pf)
-        log.info(pf)
-        bit_vector.merge_mutational_histogram_dicts(merged_mh, mh)
-    log.info(f"writing merged file: {output}")
-    pickle.dump(merged_mh, open(output, "wb"))
 
 
 if __name__ == "__main__":
